+++ pkg/diagnostics/http_tracing_test.go
@@ -36,7 +36,7 @@ import (
 	diagConsts "github.com/dapr/dapr/pkg/diagnostics/consts"
 	diagUtils "github.com/dapr/dapr/pkg/diagnostics/utils"
 	"github.com/dapr/dapr/pkg/http/endpoints"
-	"github.com/dapr/dapr/utils/responsewriter"
+	"github.com/dapr/dapr/pkg/responsewriter"
 )
 
 func TestSpanContextFromRequest(t *testing.T) {
