+++ pkg/testing/error_reader.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package streams
+package testing
 
 import (
 	"io"
