+++ tests/e2e/utils/http.go
@@ -24,6 +24,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/dapr/dapr/tests/apps/utils"
 	"golang.org/x/net/http2"
 )
 
@@ -45,7 +46,7 @@ func InitHTTPClient(allowHTTP2 bool) {
 // This should not be used except in rare circumstances. Developers should use the shared httpClient instead to re-use sockets as much as possible.
 func NewHTTPClient(allowHTTP2 bool) *http.Client {
 	// HTTP/2 is allowed only if the DAPR_TESTS_HTTP2 env var is set
-	allowHTTP2 = allowHTTP2 && IsTruthy(os.Getenv("DAPR_TESTS_HTTP2"))
+	allowHTTP2 = allowHTTP2 && utils.IsTruthy(os.Getenv("DAPR_TESTS_HTTP2"))
 
 	if allowHTTP2 {
 		return &http.Client{
